+++ packages/shadcn-auth/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @pixpilot/shadcn-auth
 
+## 0.1.54
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn-ui@0.50.0
+
+## 0.1.53
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn-ui@0.49.0
+
+## 0.1.52
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn-ui@0.48.0
+
 ## 0.1.51
 
 ### Patch Changes